Extracted SNVs are now optional outputs.
There's no point in creating a file of extracted SNVs with header lines and nothing else because it causes problems with OxoG. So now the extraction script will only preserve the file if it has non-header content. This means that everywhere else, it has to be assumed that extract_snv could return null (or an array full of `null` if called in a scatter).
And there is now a new step to filter out the `null` values from that array in the pre-processing subworfklow.